Error in aStar. float to double

+++ A Star/A Star/Node.cs
@@ -2,11 +2,11 @@
 
 class Node
 {
-    public float CSF = 0;
-    public float H { get; set; }
-    public float TEC { get; set; }
+    public double CSF = 0;
+    public double H { get; set; }
+    public double TEC { get; set; }
     public Node parentNode { get; set; }
-    public float[] location = new float[2] { 0, 0 };
+    public double[] location = new double[2] { 0, 0 };
     public string name { get; set; }
     public List<Node> connectedNodes = new List<Node>();
 }
